Use metadata-schema tooling (mdstools) for validation and migration
- Add mdstools (from the metadata-schema repository, pinned to tag 0.8.0)
  as a dev dependency; pin click<8.2 for its linkml dependency.
- Validate schemas with mdstools' validator instead of check-jsonschema:
  the schema is fetched once per run, files are validated in-process, and
  the instrument-reference check for operationParameters is included.
- SCHEMA_VERSION is now a plain release tag (0.8.0) or branch name (main)
  instead of tags/X.Y.Z / head/branch; update the version-bump workflow,
  which now also keeps the mdstools pin in sync.
- Add migrate-metadata(-dry-run) tasks (echemdb_ecdata.migrate) applying
  the mdstools migration engine to the input YAML files in literature/.
- Move the fix utilities from validate.py into echemdb_ecdata/fix.py.
- Stamp test fixtures with echemdbSchemaVersion 0.8.0.
