+++ echemdb_ecdata/entrypoint.py
@@ -48,7 +48,7 @@ from pydantic.alias_generators import to_camel
 from svgdigitizer.entrypoint import _create_bibliography
 from unitpackage.entry import Entry
 
-from echemdb_ecdata.validate import ECHEMDB_SCHEMA_VERSION
+from echemdb_ecdata.validate import SCHEMA_VERSION
 
 logger = logging.getLogger("echemdb_ecdata")
 
@@ -280,7 +280,7 @@ def convert(csv, outdir, metadata, bibliography):
 
     del metadata["dataDescription"]
 
-    metadata["echemdbSchemaVersion"] = ECHEMDB_SCHEMA_VERSION
+    metadata["echemdbSchemaVersion"] = SCHEMA_VERSION
 
     # get bibliography data and add bibdata to metadata
     if bibliography:
